Lab_5
The program is a console application for developing computer games using the "abstract factory" pattern. It allows you to create game elements, such as characters, weapons, and levels, for various game genres, such as shooters, strategy, and adventure games. Each genre is represented by its own factory that produces the corresponding game elements. After selecting a genre, the program creates game elements and displays information about them in the console. It allows you to quickly and efficiently create game elements for different types of games without the need to change the main code.